// src/characters/models/character.model.ts
export class Character {
    id: number;
    name: string;
    hitPoints: number;
    strength: number;
    defence: number;
    intelligence: number;
    class: string;
    // constructor(id: number, name: string, hitPoints: number, strength: number, defence: number, intelligence: number, heroClass: string) {
    //     this.id = id;
    //     this.name = name;
    //     this.hitPoints = hitPoints;
    //     this.strength = strength;
    //     this.defence = defence;
    //     this.intelligence = intelligence;
    //     this.class = heroClass;
    // }
}